requirements.txt
streamlit>=1.32.0
langgraph>=0.2.0
langchain-anthropic>=0.1.23
langchain-core>=0.2.0
snowflake-connector-python>=3.6.0
python-dotenv>=1.0.0
pytest>=8.0.0
